requirements.txt
# SkillVerse - Python Flask Dependencies
# Install using: pip install -r requirements.txt

# Core Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1

# Real-time Communication
Flask-SocketIO==5.5.1
python-socketio>=5.12.0

# Database
SQLAlchemy>=2.0.35
psycopg2-binary>=2.9.9

# Security & Authentication
Werkzeug==3.0.1
bcrypt==4.1.2
itsdangerous==2.1.2 # Added for security, often a dependency of Flask/Werkzeug

# Email validation
email-validator==2.1.0

# Email sending
Flask-Mail==0.9.1

# Environment variables
python-dotenv==1.0.0

# Date/Time handling
python-dateutil==2.8.2
pytz==2024.1 # Added for timezone support

# OAuth & API
Authlib==1.3.0
requests==2.31.0
Flask-Moment==1.0.5
Flask-Compress==1.14

# For future Firebase integration
# firebase-admin==6.3.0  # Uncomment when Firebase SDK is provided

gunicorn==21.2.0
